Fix infinite loop in REPLACE (Case 28181)
The REPLACE function needs to move ahead its "start index" while it's replacing substrings, so that it doesn't keep replacing the same part of the string over and over when the target string happens to contain the search string entirely (aka the case of REPLACE("http:/test.com", ":/", "://")).

The code now moves the start index to right after the replacement operation before searching again.