Make SET support calls like `SET Session.Contents(...) = ...` (Case 33958)
In this case, Session is a class that has a property called Contents. The Contents property itself doesn't have index parameters, instead the Contents of type IVariantDictionary which has a *default* indexed property.

Previously, this wouldn't work because the GenerateSetInvoker method saw Session.Contents(...) and assumed Contents was either an array member on Session or was a named indexed property on Session. To have it work before this change, the default property of Contents would have to be called out explicitly like `Set Session.Contents.Item(...) = ...`.

Now, GenerateSetInvoker will also consider that Contents is just a regular property without index parameters, and that the object it holds may have a default property with index parameters. Essentially it'll first retrieve the property value (Session.Contents), and then try to call it as a property set with parameters.

This is only supported if the named property on the object in question has a type that can be reflected upon (is known to the CLR and its properties can be queried on).

This is all very confusing - we're invoking objects' properties' properties. It only has one-level deep nesting right now because I think that's all we need to support and I didn't want to accidentally cause infinite loops.